refactor(frontend): split App into components
